TUMProject.jl Documentation

Grids

TUMProject.AbstractGrid — Type

AbstractGrid

Supertype of all grids.

TUMProject.Uniform2DGrid — Type

Uniform2DGrid <: AbstractGrid

A subtype of grid. Quadratic grid with equidistant points.

TUMProject.uniform2dgrid — Method

uniform2dgrid(h_axis::AbstractRange, v_axis::AbstractRange) -> Uniform2DGrid

Creates uniform grid using discrete uniform horzontal and vertical axis.

Laplace Operators

TUMProject.AbstractLaplaceOperator — Type

AbstractLaPlaceOperator

Supertype of all Laplace operators.

TUMProject.Laplace2DPBC — Type

Laplace2DPBC <: AbstractLaplaceOperator

Two dimensional Laplace operator with periodic boundary conditions implemented.

TUMProject.laplace2dpbc — Method

laplace2dpbc(grid::AbstractGrid) -> Laplace2DPBC

Initializes two dimensional Laplace operator with periodic boundary conditions on some grid.

Missing docstring.

Missing docstring for TUMProject.laplace2dpbc(x). Check Documenter's build log for details.

TUMProject.Laplace2DDBC — Type

Laplace2DDBC <: AbstractLaplaceOperator

Two dimensional Laplace operator with dirichlet boundary conditions implemented.

TUMProject.laplace2ddbc — Method

laplace2ddbc(grid::AbstractGrid) -> Laplace2DDBC

Initializes two dimensional Laplace operator with dirichlet boundary conditions on some grid.

Missing docstring.

Missing docstring for TUMProject.laplace2ddbc(x). Check Documenter's build log for details.

Reaction Diffusion Systems

TUMProject.AbstractReactionDiffusionSystem — Type

AbstractReactionDiffusionSystem{T<:AbstractFloat}

Supertype of all reaction diffusion systems.

TUMProject.FastReactionSystem — Type

FastReactionSystem{L<:AbstractLaplaceOperator,T<:AbstractFloat}  <: AbstractReactionDiffusionSystem{T}

Fast-Reaction system type.

TUMProject.fastreactionsystem — Method

  fastreactionsystem(Δ, f, g, D_u, D_v, eps) -> FastReactionSystem

A subtype of a reaction diffusion system. It is a system of the form

\[\begin{aligned} \partial_t u = D_u \Delta(u) + \varepsilon^{-1}f(u,v) \\ \partial_t v = D_v \Delta(v) + g(u,v) \end{aligned}\]

Missing docstring.

Missing docstring for TUMProject.fastreactionsystem(dS, S, N, t). Check Documenter's build log for details.

Limit Systems

TUMProject.AbstractLimitSystem — Type

AbstractLimitSystem{T<:AbstractFloat}

Supertype of all limit systems.

TUMProject.FastReactionLimitSystem — Type

FastReactionLimitSystem{T}  <: AbstractLimitSystem{T}

A subtype of a limit system corresponding to a fast-reaction system.

TUMProject.fastreactionlimitsystem — Method

  fastreactionlimitsystem(Δ, h, g, D_v) -> FastReactionLimitSystem

The limit system of a fast-reaction equation. It is a system of the form

\[\begin{aligned} 0 = f(u,v) \\ \partial_t v = D_v \Delta(v) + g(h(v),v) \end{aligned}\]

In this form it is assumed that one can solve f=0 for u = h(v), so that the system is only depending on v.

Missing docstring.

Missing docstring for TUMProject.fastreactionlimitsystem(dS, S, N, t). Check Documenter's build log for details.

Comparing

```@docs TUMProject.plotinitialconditions(x, y, xname, yname) TUMProject.animatesystem(x, y, xname, yname, ts, limit, name, fps) TUMProject.plotcomparenorms(sol, limitsol, sol2, limitsol2, p, ts, figlabel, figlabel2)